Refactor code into packages

more refactoring
